manifests: push all mutations on objectupdate
the manifests package should perform only minimal mutation,
and only if/where needed.

RTE daemonset was an exception for legacy reasons, not
by deliberate design. At long last, move mutations (and their
tests) in the objectupdate/rte package.
